Building a Smarter AI-Powered Spam Classifier
In this tutorial, learn how to build an AI text classifier and deploy it as a microservice to be embedded into your own application. The tutorial uses a data set that describes animals and builds a classifier to identify the sentences that describe their habitats. You can easily apply this same approach to different data sets for your own use cases.